requirements.txt
edgedb
requests
python-dotenv==1.0.1
flask
discord-oauth2.py